+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.0 (2026-05-29)
+
+
+### Features
+
+* add mac os app signing to release pipeline
+
+### Bug Fixes
+
+* fix tray misclick and autostart for each platform
 ## 1.0.0 (2026-05-01)